% Auto-generated: do not edit by hand
\name{''DataTable}

\alias{''DataTable}

\title{DataTable component}

\description{

}

\usage{
''DataTable(id=NULL, bordered=NULL, box_selected_keys=NULL,
box_type=NULL, col_filtered=NULL, data=NULL,
filterDropdownVisible=NULL, footer=NULL, locale=NULL,
pagination=NULL, row_filtered=NULL, scroll=NULL,
searchText=NULL, selectedcell=NULL, showHeader=NULL,
size=NULL, style=NULL, title=NULL)
}

\arguments{
\item{id}{Character. }

\item{bordered}{Logical. }

\item{box_selected_keys}{Unnamed list. }

\item{box_type}{Character. }

\item{col_filtered}{Unnamed list. }

\item{data}{Named list. }

\item{filterDropdownVisible}{Unnamed list. }

\item{footer}{Character. }

\item{locale}{Named list. }

\item{pagination}{Named list. }

\item{row_filtered}{Unnamed list. }

\item{scroll}{Named list. }

\item{searchText}{Unnamed list. }

\item{selectedcell}{Named list. }

\item{showHeader}{Logical. }

\item{size}{Character. }

\item{style}{Named list. }

\item{title}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

