% Auto-generated: do not edit by hand
\name{''Network}

\alias{''Network}

\title{Network component}

\description{

}

\usage{
''Network(id=NULL, data=NULL, event=NULL, event0=NULL, event1=NULL,
event2=NULL, event3=NULL, event4=NULL, event5=NULL,
event6=NULL, event7=NULL, event8=NULL, event9=NULL,
fit=NULL, focus=NULL, moveTo=NULL, options=NULL, run=NULL,
selection=NULL, style=NULL)
}

\arguments{
\item{id}{Character. }

\item{data}{Named list. }

\item{event}{Named list. Using 'setProps' to set event props}

\item{event0}{Named list. }

\item{event1}{Named list. }

\item{event2}{Named list. }

\item{event3}{Named list. }

\item{event4}{Named list. }

\item{event5}{Named list. }

\item{event6}{Named list. }

\item{event7}{Named list. }

\item{event8}{Named list. }

\item{event9}{Named list. }

\item{fit}{Named list. }

\item{focus}{Named list. }

\item{moveTo}{Named list. }

\item{options}{Named list. }

\item{run}{Character. run your javascript here}

\item{selection}{Named list. }

\item{style}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

